mediav2: Limit checks based on flags to Android V and above
AdaptivePlaybackTest.java has tests that are to be enabled,
based on a flag. This check that uses flags needs to be limited to
Android V and above as these tests are run in MTS on older versions.

Bug: 332574755
Test: Ensure there are no unexpected failures with the following on
Android R device.
atest CtsMediaV2TestCases:AdaptivePlaybackTest -- --module-arg\
CtsMediaV2TestCases:instrumentation-arg:codec-prefix:=c2.android.

Change-Id: I2252755e336fd5d931bd7b38be2a7899594afbc4