Modify displayPoint calculation to adapt displays with widgets cutout
When calculate the x value of displayPoint on displays with widgets
cutout may wrongly acquire the full size of display. Similar to
calculate the y value of it needing to consider the cutout of camera,
we also consider the widgets cutout when calculate x value.

Two kinds of circumstances should be concider:
1. The widgets area covered status bar area
The x value calculated with the min-insets.
2. The widgets area not covered status bar area
The x value calculated with the cutout-insets.

Bug: 333027516
Test: AssistantContentViewTest#testAssistantContentViewDimens

Change-Id: I015843f707393bd0a33930f92ba7e38efe7e2e75
Signed-off-by: Duo Huang <huangduo@xiaomi.corp-partner.google.com>