Adopt shell permission identity in testSendCallEvent CTS
Following the recent mainline changes, CTS has been failing for
TransactionalApisTest#testSendCallEvent because the required permissions
have not been granted. The prior logic would've skipped the query to
UserManager#isUserAdmin but now that it has been moved out of the if
block, this API is always called, which requires either the
MANAGE_USERS, CREATE_USERS, or QUERY_USERS permission. Adopting the
shell permission identity during CTS should resolve this.

Bug: 330184947
Test: atest TransactionalApisTest
Change-Id: I0d38b4b576dc9daf832b0d2db7fcf2b015732b41