enable the cached callbacks flag when running Telecom CTS
Recently, a fix was merged that cached callbacks that were being executed
before the servcie was set.  That change eliminated all flake for the
CallEndpoint tests locally. However, the flag has not been enabled in
most test branches and is still failing. It could take weeks before
the flag is enabled and the flake is fixed on every branch.  In  order
to put those tests back into compliance, the flag should be enabled in
CTS.

Fixes: 321401907 (available endpoints flake)
Fixes: 332810069 (get current endpoint flake)
Bug: 321204022 (Legacy Telecom CTS module flake)
Bug: 328092080 (CUJ module flake)
Test: existing CTS modules
Change-Id: I26b6e20ef2154f153588bbfb4db7bcf8e144bbb1