Protect against stopping dream session after dream destroy.
This change adds measures to prevent crashing when a test attempts
to stop a dream session after the dream has gone away and the
binder is no longer active. The first measure is to track the
service and clean references on death. The second is to catch
RemoteExceptions from interacting with remote objects during
cleanup.

Test: atest DreamServiceTest
Fixes: 333766297
Change-Id: Ib070a93938af5df86a66b83749b53db56d785dae