InputAtomsTest: reformat
Every time I've uploaded a CL to this file ktlint has complained about
formatting issues, mostly on lines I didn't touch. This leads to an
annoying process of selectively reverting the changes it made to
untouched lines and then adding --no-verify to my repo upload command.

Avoid this by reformatting the file to comply with its guidelines. At
the same time, standardize on a single indentation level increase for
function call arguments, as previously parts of the file used two. (The
style guide [0] is not particularly clear on this point, but if one
interprets "block-like construct" to include argument lists then one
level increase is correct. Besides, it makes the lists of matchers more
readable because fewer have to be wrapped.)

[0]: https://developer.android.com/kotlin/style-guide#indentation

Bug: 245989146
Test: check ktlint pre-upload hook doesn't complain
Test: build InputAtomsTest
Flag: EXEMPT whitespace change
Change-Id: Ia75b468e542b5c286bfeea87576ea9e2afedd14b