Prevent activity recreation of ActivityEmbeddingFocusTests
The CL prevents activity recreation by adding the configuration types
to android:configChanges, so that primaryActivity and secondaryActivity
will both point to the expected instances instead of the stale ones.

This CL also waits for the potential change of the window focus to be
settled since the key events are sent in a row programmatically. This
can make the test more stable and closer to the use cases in the real
world. i.e., We don't want the keys to be sent to the old focused
window.

The last change of the CL is to move the accessing of the activities to
the main thread, to prevent unexpected race conditions.

Fix: 334246829
Fix: 333637163
Fix: 347823480
Flag: NA
Test: atest ActivityEmbeddingFocusTests
Merged-In: I26044073a96e27761b8c90485bbecbd784129bdd
Change-Id: I26044073a96e27761b8c90485bbecbd784129bdd