Correctly measure available free memory in CarWatchdog CTS
Refactored the writeToFos method to correctly measure the total
available free memory in the process. Runtime.freeMemory() method
returns the currently allocated memory that is free, but does not
consider the available unallocated memory. If a process where to exceed
the currently freeMemory, then the JVM would simply allocate more memory
to the process (without exceeding the Runtime.maxMemory() amount).

Test: atest --rerun-until-failure 5 CtsCarTestCases
Bug: 335875865
Bug: 329680291
Change-Id: I30a706c32810698b9745ec5c9ea07aab1bb53056