CodecEncoderSurfaceTest: Add a note for persistent surface issues in MTS
CodecEncoderSurfaceTest was updated to test persistent surface mode in
Android U. Though this is expected to work for older versions, there are
some devices where s/w encoders do not work as expected with h/w decoders
in persistent surface mode. These show up as failures in MTS mode.

For now, have added a note explaining this behavior in the test.

Bug: 306103736
Test: Review

Change-Id: Iee51108740d06390ac4cf958f39fb194e813b9b8