Only set PhoneCapability#maxActiveVoice for V+
When setting up the default phone capability, this CL now checks that the android version is V+ before setting the maxActiveVoice field to a default field. This resolves an error where we were trying to set maxActiveVoice field on devices with an old HAL that didnt have the field maxActiveVoices.

Test: build + atest SimultaneousCallingRestrictionsTest
Fixes: 335386530
Change-Id: Ica1fd51d6ccb5c50b252f43b802c27eeaf3e80d4