Enable --secondary_user_on_secondary_display for CtsKeystorePerformanceTestCases CtsEffectTestCases CtsSliceTestCases
--secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 337339189
Bug: 337340953
Bug: 337341742
Test: atest --user-type secondary_user_on_secondary_display CtsKeystorePerformanceTestCases CtsEffectTestCases CtsSliceTestCases
Change-Id: I9be8c4144d0d39242b24e429e94941b6a6201f57