Enable --secondary_user_on_secondary_display for CtsFileSystemTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 338160385
Test: atest --user-type secondary_user_on_secondary_display
CtsFileSystemTestCases
Test: atest CtsFileSystemTestCases

Change-Id: I544a00c8d3e346544199bb727dc5c1e07a357adc