CarOccupantZoneManagerTest#testGetSupportedInputTypes_validatedInputTypes failing due to unavailable display
Occupant zone mapping was added for the display compat display. Since
that display is created dynamically, the display has not yet been
created for the display when the test is run due to which it is not an
active config and has no supported input types yet.

Only run on the public display configs since none of them are created
dynamically.

Bug: 337384226
Test: atest CtsCarTestCases:android.car.cts.CarOccupantZoneManagerTest
Change-Id: Ia82b69575b1d8625303eacc3d64c6443c33b18dd