Make MediaProjectionActivity's View Finding More Robust
We're now adding an additional way of looking for a UiObject within the
activity by using the UiAutomatorUtils. This is helpful in cases where
the previous approach (i.e. relying on the UiScrollable#scrollIntoView)
fails to find the UiObject.

Note that this change is safe w.r.t regressions to already passing
tests, since we are only adding an additional way of finding a UiObject
after already exhausting the previous approaches.

Bug: 333476576
Test: atest MediaProjectionTest
Change-Id: I8a66e0f662cbb707bba1ad68f64b74f622fe4299