ITS: use dataclass for zoom test data
Details: using a dictionary to store data may not be ideal, especially
when it's shared between modules. Changed behavior to use a dataclass,
which allows for data to be set with keyword arguments. Also, accessing
values is much easier (no longer need to look up what key name was
used).

Bug: 338638040
Test: ran zoom tests locally, same behavior as before.
Change-Id: Ic33425197c67de3b39e8c9418408fe75df880d86