Update the "shouldTestTelecom" requirements.
Over time we've updated `shouldTestTelecom` to look for whatever the most
current set of feature flags is.  This is actually not correct as a device
will have a vendor SDK level set at the time it is built.  This means that
a device will be locked at a particular set of feature flags based on the
SDK level the device was originally made for.  When CTS recertification
takes place on OS upgrades, the CTS test suite for the most current OS
version can be running on a device that still declares feature flags for
an older SDK version.
Updated the feature flags checks to ensure that we are running the Telecom
CTS tests on all devices that actually support Telecom when they were
made.

Fixes: 339300693
Test: This is a CTS test fix.
Change-Id: I7ee61c7aae466e404f1788c01a679816ffa29871