Correct CTS to reflect new behavior on Dialog
We had a report of a flake caused by the new behavior of the
Escape key, as it will now dismiss dialogs even if they are not
cancelable. This CL also adds a new test to ensure that the
updated behavior when the dialog is set to close when touching
outside is respected.

Bug: 333416590
Bug: 332717129
Test: ABTD
Flag: TEST_ONLY
Change-Id: Ie0465d99c75966cca124ebff6f963d5f4e07bafd