Skip checking activity bounds for devices with default multi-window
There are devices that always launch activity in multi-window. On those
devices, the activity bounds may come after display bounds change,
because the task bounds is controlled by sysUI.

Skip the activity bounds change for those devices, and add a separate
test case to ensure the display bounds are still up-to-date.

Fix: 339104284
Test: atest android.server.wm.activity.ConfigurationCallbacksTest
Change-Id: I4bb97f7968e1cd55bf064d122c505ff332cefd25