Increase timeout for JvmtiHostTest990 test
On some devices this is known to take a long time. Increasing the
timeout as it started failing with the current timeout. Earlier it used
to run close to the timeout and now it is a bit over the current
timeout. So increasing the timeout as a temporary fix. We need to
investigate on why the test is taking so long.

Bug: 336711414
Test: cts -m CtsJvmtiRunTest990HostTestCases.android.jvmti.cts.JvmtiHostTest990
Change-Id: I48f370c350e2c0fb662ea302b42d66461c124353