mediav2: Do not queue CSD to writeSampleData in Muxer API
Output from the encoder that is flagged as CodecConfigData can not
be queued to MediaMuxer's writeSampleData.

MediaMuxer doesn't support data flagged as CODEC_CONFIG in
writeSampleData(). For VP9 which doesn't have start codes, sending
the CODEC_CONFIG data in writeSampleData leads to this data being
stored as a frame and then the resulting file has issues when being
decoded.

Bug: 332052663
Test: atest CtsMediaV2TestCases

Change-Id: Ic6ea7f594814cc2dced7fce0416e1e2b9be1d783