Update tests for asl-in-apk
Test that the app metadata can be extracted from the apk as a raw
resource and check that the file will not extracted if it is larger than
the app metadata file size limit.

Refactored some existing tests to included subsequent updates with and
without app metadata and removed some redundant tests.

Bug: 338418504
Test: atest InstallAppMetadataTest
Flag: NONE
Change-Id: I70eaa571dab687d9fbd95583aa3b76f3f495f4b4