Run libcore tests in mts mode by default
It's needed because test/mts/tools/mts-tradefed/res/config/mts.xml
will not be used in the future when MCTS is fully ready. The mts.xml
tells the CoreTestModeFilter that CtsLibcoreTestCases
and CtsLibcoreOjTestCases should run in the mts mode.

Also, the default mode change here ensures that the same set of
test cases run in various CI as in the MCTS test runs.

Bug: 340212375
Bug: 287042293
Test: atest CtsLibcoreTestCases CtsLibcoreOjTestCases
Change-Id: I40e86af07aa3381f84b81ab86bdf7322e24c8d86