mediav2 CTS: Reduce runtime of CodecEncoderSurfaceTest
Reduce the runtime for CodecEncoderSurfaceTest and VideoEncoderTest.
This change reduces runtime from ~19 min to ~9 min for
CodecEncoderSurfaceTest and from ~4 min to ~2 min for
VideoEncoderTest. This is done by limiting the frame
limit so that it runs only for 30 frames if frameRate
is less than 30 fps otherwise it runs for 1 sec duration
of the clip.

Bug: 339872733
Test: atest android.mediav2.cts.VideoEncoderTest
Test: atest android.mediav2.cts.CodecEncoderSurfaceTest
Change-Id: I7d59e18ab81eba0164f6fbf8e91a0e5d62ec2c0b