Fixing CTS Verifier USB Companion echoing zero size package
The accessory test activity was echoing max size buffers after overwriting
the first byte with the zero sized package received at the end of a
transfer which is a multiple of the package size, this was causing a
single byte to be mismatched with the source on the CtsVerifier USB
Accessory Test. Reading that byte into a separate buffer and ignoring it
for now.

Fixes: 340999930
Test: Manually tested
Change-Id: I9fdbd421137f7ee2e2afcea619e50cdcd24b6e9a