Adjust BrightnessTest
Brightness events can be inconsistent because of the asynchronous nature
of the brightness control in the test as well as the uncontrollable
parts of the tests (live ALS).

The test is altered so that instead of getting the latest event and
verifying it if what we think, we query for events until we find the one
that we are looking for.

Test: atest BrightnessTest
Bug: 340878922
Change-Id: Ie1ad819fb18270e9483578f06cff108ec4e2906d