VIC finalization: baseline new NewApi errors
Now that VIC is 35, the errorprone build will detect access to APIs
introduced in 35 that are not guarded by the proper INT_SDK check.

To unblock the last steps of the VIC finalization, baseline these
errors for now. The team owning this app is responsible for keeping the
baseline, or adding the necessary checks.

Bug: 335304236
Test: treehugger (errorprone target)
Change-Id: Ibd5783ef1f83819769e3f55211249f76c6d96aa1