Opt out force-queryable when installing test IMEs
This is a preparation to add a regression test for Bug 342105635.

By default <target_preparer> installs APKs with --force-queryable
flag, which effectively makes APKs visible from any other packages
even without <queries> elements in the observers'
AndroidManifest.xml. While such a behavior is useful for causal
testing purposes, it basically makes it impossible for us to test
package visibility conformance in CTS.

With this CL test IMEs will be installed without force-queyable
mode. As long as there is no test breakage, this CL should help us
test things in a way that is closer to end user environment.

Bug: 342105635
Test: atest CtsInputMethodTestCases
Change-Id: I8bbb24b561695e76323d8f19d376c8782a674601