Adding BAL test including system activities
Add a CTS test to the ActivitySecurityModelTest that checks whether
background launches are allowed on top of a system activity, depending
on the value of the flag.

The test starts an activity for handling the
"MANAGE_UNKNOWN_APP_SOURCES" action. Because the exact activity name
that'll be launched is unknown (and may be an internal, un-exported
activity), the test waits for the launching activity to lose focus and
captures the activity that does get launched.

Subsequent assertions check that this is the activity at the top of the
task stack, or the one sandwiched between two test activities.

Bug: 339403750
Test: atest CtsWindowManagerBackgroundActivityTestCases:android.server.wm.ActivitySecurityModelTest
Flag: android.security.asm_opt_system_into_enforcement
Change-Id: I6d74f02de5da337c31071a82deaad47f6214afb8