Enable secondary_user_on_secondary_display for CtsFragmentTestCases
secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 342520135
Test: atest CtsFragmentTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsFragmentTestCases
Change-Id: Ibb33ac7aafc3e87209bdd30e1399655e242240f8