Calculate expected values before activity launch in ManifestLayoutTests
`testLayout()` in `ManifestLayoutTests` compares some expected layout
values with the actual layout values of an activity. However, these
expected values are calculated after the activity launch and not
before. This means that the activity layout and the expected layout
could potentially be calculated on different display states.

An OEM may have a persistent navigation bar when a freeform window is launched. To account for these insets when calculating the size and position of a freeform window, the expected layout should be calculated on the same display state as the activity is launched on. So we should first calculate the expected layout (which doesn't effect the display state) and then launch the activity.

Fixes: 341003131
Test: atest CtsWindowManagerDeviceAnimations:ManifestLayoutTests
Change-Id: I03ff4894853ae83e47968585e3ada843cd4ba767