Enable secondary_user_on_secondary_display for CtsVideoTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 342506458

Test: atest --user-type secondary_user_on_secondary_display \
          CtsVideoTestCases
Test: atest CtsVideoTestCases

Change-Id: Ifb538ad560be248bfc1ba4be8d71f64b6f2a0278