Enable secondary_user_on_secondary_display for CtsMediaEditingTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 342433321

Test: atest --user-type secondary_user_on_secondary_display \
          CtsMediaEditingTestCases
Test: atest CtsMediaEditingTestCases

Change-Id: Id8f42cd6b541313c84eba7d9e4818c2801d106d4