Fix testZeroLengthSubElementIEOnMockWifi fail
The current design for updating scan result cache is based by scan request.
So we can see this test fail when wifi is off before running the test.
It is because mocked scan result didn't be updated since there is no more
scan request in test case. (If wifi is connected before running the test,
the test is passed since scan request is triggered in setUp method and it can
cause mocked scan result is updated)

So we should add a scan request in the test case to cover all condition.

Bug: 343935834
Bug: 342518912
Test: atest -c android.net.wifi.mockwifi.cts.MockWifiTest#testZeroLengthSubElementIEOnMockWifi
Change-Id: I05eed60644f42eb40c8b02549eaa345081988256