Add more GWP-ASan tests (incl. recoverable tests)
I accidentally deleted the recoverable tests in aosp/3033333. Oops.

While I was re-adding them, I also added a new test that ensures that
recoverable gwp-asan works as expected for both the
"gwpAsanMode='always'" and 'gwpAsanMode='default'" (or unspecified). It
should be in recoverable mode both times, it's just whether process
sampling is enabled or not.

Bug: N/A
Test: atest CtsGwpAsanTestCases
Change-Id: Ia620c0d476221480a309884750ee8f8f75a1f825