Split CtsLibcoreTestCases at runtime with test_module_config
This CL doesn't affect the tests, it only creates the new modules.
A CL (aosp/3108997) in libcore/tests will change all the TEST_MAPPING files to use
these targets.
That CL ran the presubmit group in atest before and after the changes to ensure the
number of tests ran still matches.

Test: mm
Test: see aosp/3108997
Change-Id: I1d289c51525c0a5aed07bf5d0fe5d73cef0be20d