Allow activity start in tests run in sandbox mode
Starting activities in the SDK sandbox are normally not allowed.
However, for the purposes of CTS testing, they can be temporarily
enabled to allow the test to proceed. This requires a
START_ACTIVITIES_FROM_SDK_SANDBOX permission. Modify any CTS tests that
fail when run in the sandbox mode to allow the activity to be started.

Additionally, mark the activity as exported so that the sandbox process
can see this activity.

Bug: 323156668
Test: atest <test-name> -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Change-Id: I690ebec77100eb2e394d57d17fcc7b6ddebf8ed8