Enable --secondary_user_on_secondary_display for CtsTextClassifierTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Bug: 344042127
Test: tested on gcar_md emu
Test: atest --user-type secondary_user_on_secondary_display CtsTextClassifierTestCases
Test: atest --user-type secondary_user CtsTextClassifierTestCases
Test: atest CtsTextClassifierTestCases
Change-Id: I3b2169f0c6bf3339e2613fdd0f350bdd4749deeb