Fix ContentResolverTest#testUpdate for the sandbox
When the test is run in the SDK sandbox mode, the test runs in the SDK
sandbox process while the ContentProvider is run in the app process.
When calling ContentResolver.update() with null values in the sandbox,
it goes through ContentProviderNative as part of the IPC which throws a
NullPointerException. This code flow does not occur when both the test
and the ContentProvider run in the app and a different exception is
thrown.

Modify the test such that the correct exception is checked while running
the test in sandbox mode.

Bug: 344539972
Test: atest ContentResolverTest#testUpdate -- --enable-optional-parameterization
--enable-parameterized-modules --module-parameter run_on_sdk_sandboxFix
ContentResolverTest#testUpdate for the sandbox

Change-Id: I34c2c760b3fbf40ceb2c2591f3c4b2a48790b7d6