fix(magnification): sleep when teardown to wait for async cleanup
In AccessibilityMagnificationTest, first we increase the timeouts to
improve the cuttle fish devices' test stability. Besides, in teardown step
we add a thread sleep after disabling service. The purpose is to let the
async background service cleanup process, like magnification sysui
connection releasing, finish in the teardown scope, so that it will not
affect the next test iteration.

Bug: 342631569
Flag: EXEMPT bugfix
Test: atest AccessiblityMagnificationTest
Change-Id: I156e44ec1ae7311fe3bc4dc13e7dc3b7227a55ed