Enable secondary_user_on_secondary_display for CtsMediaCUJSmallTest.
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

PINCH_TO_ZOOM tests have been modified to create events on the correct
touchable region according to the activity's display.

DEVICE_LOCK tests are being skipped for the
secondary_user_on_secondary_display case because we currently do not
have a way to send a "press the sleep button" event to a secondary
display.

Bug: 342432386

Test: atest CtsMediaCUJSmallTest
Test: atest --user-type secondary_user CtsMediaCUJSmallTest
Test: atest --user-type secondary_user_on_secondary_display \
        CtsMediaCUJSmallTest

Change-Id: Icf36a5cb6266d22f979bb1dbaf03ec322f102eed