Enable secondary_user_on_secondary_display for CtsMidiTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 344072417
Test: atest CtsMidiTestCases
Test: atest --user-type secondary_user CtsMidiTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsMidiTestCases
Change-Id: Id8c9b987faceb5f25caecb428e9c2716239c07d0