Ensure the test doesn not have permission before running
Previously, this test relied on the absence of CREATE/QUERY/MANAGE user permissions to trigger a security exception. This behavior was not explicitly validated, creating a potential risk of false positives if unexpected permissions were granted.

Added @EnsureDoesNotHavePermission({CREATE_USERS, QUERY_USERS, MANAGE_USERS}) annotation to proactively verify the absence of required permissions before executing the test, ensuring a consistent and reliable testing environment.

Also refactored tests to use granular assertions for better error reporting. Previously, we only checked for the SecurityException. Now, assertions confirm both the presence of an exception and that it is specifically a security exception and log messages.

Test:  atest android.multiuser.cts.UserManagerTest -c
Bug: 335116158
Flags: Exempting flags as it's only a test fix

Change-Id: I76b9302b043058ace77534d75559caa1d4d4498c