Finish the showWhenLockedActivity when the lockscreen session is ending
Due to a launch root task being present, the showWhenLockedActivity gets
launched in the launch root task. Due to this, when the keyguard comes
on, this activity remains in the visible state but is behind the
keyguard due to it being attached to the launch root task. This results
in input events not being received since the surface is not visible.
Instead, manually finish the activity until the behavior for
setShowWhenLockedActivity has been decided with the presence of a launch
root task.

Bug: 322917959

Test: atest CtsWindowManagerDeviceActivity:ActivityLifecycleTopResumedStateTests
Change-Id: I9ed92815c6f8fa232358d18ae231dac1557cdf8f