mediacuj CTS: Override ring volume for call notification test
If the ring volume of the device is muted, then the playback continues
even when an incoming call is placed and the playback suppression reason
does not change. Thus, we are adding a change that checks if the ring
volume is 0 and overrides the ring volume to the volume index 1.

Bug: 346224792
Test: atest CtsMediaCUJSmallTest
Change-Id: I7378ea8bb09135ce81de16a8f44c7ac84a0ef1eb