Make DeviceAndProfileOwner apk backwards compatible
The new headless-device-owner-mode="single_user" attribute introduced
in Android V isn't backwards-compatible when the APK is running on
Android U. As it turns out, we do run some CTS tests using this APK
on Android U devices for mainline MTS, so we need to make them
backwards-compatible. Fix this by forking the device_admin.xml to
only include the new attribute value on devices running on SDK leve 35
or above.

Bug: 346246743
Test: MixedManagedProfileOwnerTest on Android U
Change-Id: I4e72c4af948f4ea63c1e97a8d7e50788a4850f3f