Add MD support for CtsSystemIntentTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
current user) interacting with the driver's display.

Bug: 343812358
Test: atest CtsSystemIntentTestCases && atest --user-type secondary_user_on_secondary_display CtsSystemIntentTestCases
Flag: TEST_ONLY
Change-Id: Ifcd89df20ee2717bba347165bdfcb752f859dc95