Revert^2 "Fix BlurTests assumptions about insets"
This reverts commit 16a8f40fd3b3f3ef1b2f2243250a068021ee4499.
This relands commit 575eb58964ae23dda226aa45bbd44aab24de6016.

The difference from the original CL is that we remove some lines
that were "compensating" for the offset of the screenshot bounds,
which is no longer needed because the screenshot bounds now cover
the whole screen.

Reason for revert: Test failure fixed
Bug: 348668718
Bug: 347853456
Test: atest CtsWindowManagerDeviceAnimations
Flag: EXEMPT bugfix to test

Change-Id: I8d4173b2a30dcd90c0673b6dd6025a8f7e0c0981