Add test to ensure LegacyKeystore deprecation
This adds a test to ensure we get the right response when the flag `FLAG_DISABLE_LEGACY_KEYSTORE_GET` is disabled for the get and put function and when the get flag is enabled we return a error. There will be a later cl to include a test for enabling the `FLAG_DISABLE_LEGACY_KEYSTORE_PUT_V2` flag.

Test: atest CtsKeystoreTestCases:LegacyKeystoreTest
Bug: 307460850
Change-Id: I4d51405f096a479cd88a2a1f748a5fe12d98cdbc