Camera: Stay away from 1.0x for zoomSettingsOverride tests
Current camera API behavior is that if CaptureRequest.CONTROL_ZOOM_RATIO
is 1.0, the framework makes sure the CaptureResult.CONTROL_ZOOM_RATIO
is 1.0, and the actual zoom is reflected in
CaptureResult.SCALER_CROP_REGION. This throws off the look-up logic
between requests and results.

Add a tiny offset to 1.0x to address the test issue. Filed b/350076823
to fix the root cause.

Flag: TEST_ONLY
Test: Run PerformanceTest on Pixel; vendor testing
Bug: 349743475
Change-Id: I19c00dadf13529831e94b7539da6c3aadafc77e4