Fix failed ConcurrentMultiUserTest
Sometimes when Trade Federation creates a new user and starts it,
the user might not be initialized but it is sufficient to run the
test. If we only look for initialized users, we won't find the
user we want. So this CL removed user initialization requirement
when looking for concurrent users.

Fixes: 350560607
Test: atest com.android.server.inputmethod.multisessiontest.ConcurrentMultiUserTest
Change-Id: I8a20ef244b1c8bb8e66fb91becc5fa9adaecc39d
Flag: TEST_ONLY