Enable secondary_user_on_secondary_display for CtsTransitionTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 349712614
Test: atest CtsTransitionTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsTransitionTestCases
Change-Id: I998baf435cad4017e67c8056bdb19df61baefdc9