Use RAW stream for 3a in test_in_sensor_zoom
At certain zoom ratio's different physical cameras
may be used for 3a. This may not be the case for RAW capture
since logical cameras may not change the underlying physical
camera when RAW streams are configured. Therefore, use a RAW
stream in the 3A session for consistency of settings.

Also use zoom ratio 1.0 for do_3a() while capturing RAW images
at different zoom ratios (with CROPPED_RAW). This is so that the
3A settings calculated by the camera HAL have the same area under
consideration as the original RAW image (zoom ratio 1.0) being used to
compare against. Otherwise, even with the same lens there can be a
difference in brightness. The reason behind not doing this for
test_zoom.py is that in that test we test for circle size and center
coordinates, which aren't really affected by brightness. In
test_in_sensor_zoom.py, we are comparing patches of the image directly
since RAW_CROP_REGION may not be exactly centered.

Bug: 349804199

Test: Vendor Testing

Flag: TEST_ONLY

Change-Id: If7333a6f9383b2259e9c4687f4b416f78690296a
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>