Update CTS test for Android Security b/283103220
On devices which do not support picture-in-picture mode, the target APIs ('setPictureInPictureParams()' and 'enterPictureInPictureMode()') return early, without reaching the fix code. Due to this, even in the presence of the fix, the target APIs do not throw any exception, resulting in test failure.

Updated the test to skip the test on devices which do not support picture-in-picture mode.

Bug: 343534667
Bug: 350364180
Test: Ran the testcase on android-12.0.0.r28 with/without patch

Change-Id: I4569bf6333051faafe42d99f277f50a328140b44