CtsWindowManagerDeviceTestCases-7 failures
Description:CtsWindowManagerDeviceTestCases are failing on devices
with a resolution of 480x768 and a density of 320 (sw240dp), as the
UI element is not sufficiently visible for drag and drop operations.
Solution:changed the margins of the file drag_drop_layout.xml with
respective to above screen width device.
Bug: 351287048
Below are the failures under
android.server.wm.DragDropCompatTest:
1.testOverNowhere
2.testNoExtraEvents
3.testEnterExit
4.testDrawableState
5.testDragShadowWhenPerformDrag
6.testAcceptingGroupInTheMiddle
7.testBlackHole

Test: run cts -m CtsWindowManagerDeviceTestCases

Change-Id: Ib57497fad41a7d73517d69e3b0ae18d377cd9df1