CameraITS: Make build/envsetup.sh zsh compliant
zsh uses $path as an alias to the enviroment variable $PATH.
The envsetup.sh script sets $path when looking for
libtinfo.so.6 which currently also resets $PATH when used
with zsh shell which is less than ideal.

To prevent zsh environment from breaking, this CL changes $path
to $python_path in envsetup.sh.

Bug: n/a
Test: Manually verified that zsh is functional after
      source build/envsetup.sh
Flag: TEST_ONLY
Change-Id: Ie611656aa55717fcea61825ea1d14b75c5c6e6e4