Fix ISDBT frontend test error when calling createFrontendSettings
The TunerTest.createFrontendSettings(FrontendInfo) method called
TunerVersionChecker.isHigherOrEqualVersionTo(int) when the frontend type
is ISDBT. The isHigherOrEqualVersionTo method is annotated as @TestApi
which can only be accessed by method annotated as @Test. Thus the test
failed with method not found.

To fix this issue, we can simply inline the logic to avoid using TestApi
in the static utility function.

Bug: 349530943
Test: Run TunerTest#testResourceReclaimedDifferentProcess on partner device
Change-Id: I804c4795559fddf24b8d33c4afef3b42b151eff7