DO NOT MERGE Allow for CAR to disregard Immersive fullscreen test.
Android Auto has the option to control immersive fullscreen
through setting config_remoteInsetsControllerControlsSystemBars.
If this is set to True the test "testAddingImmersiveWindow" will
fail as does not expect this behaviour.

This test was on previous cookies blindly disregared for Auto but
supposedly fix in generic way in SHA1 2ce6844but but not sufficient.

Test: atest
CtsWindowManagerDeviceTestCases: LayoutTests#testAddingImmersiveWindow

Bug: 350908653
Change-Id: Ia0056fc7ae2d94efc1745de5dc8929746c200a1d