Allow activity start for acceleration tests run in sandbox mode
See go/sdk-sandbox-cts-tag and go/cts-in-sandbox-how for more details
about running tests in the SDK sandbox mode (does not affect tests
running in regular app mode).

Starting activities in the SDK sandbox are normally not allowed.
However, for the purposes of CTS testing, they can be temporarily
enabled to allow the test to proceed. This requires a
START_ACTIVITIES_FROM_SDK_SANDBOX permission. Modify any CTS tests that
fail when run in the sandbox mode to allow the activity to be started.

Additionally, mark the activity as exported so that the sandbox process
can see this activity.

Bug: 323156668
Test: atest <test-name> -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Flag: TEST_ONLY
Change-Id: I7781f33f00b7a56fd3cf3f1aef6f5e74aafebe48