Fix using instrumentation args in interactive
When a test is marked with @Interactive annotation, it calls ConnectViaAdbToHostStep as part of applyAnnotation setup. That step does not have its own automation class, so running a test target without the ENABLE_MANUAL flag will fail on every single one of them.

Fixed by checking a value of that step prior to running automation or demanding manual interaction. Works when, for example, a device is plugged already and passes AdbToHostStep check right away.

Bug: b/352485189
Flag: TEST_ONLY

Change-Id: Idfe4c15dfca2994a6aab5e8e3a30b44129ff0315