Enable CTS test for secondary_user_on_secondary_display in bulk
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Tested on gcar_md emu

Bug: 354233217
Test: atest --user-type secondary_user_on_secondary_display CtsLibnativehelperTestCases CtsNativeMediaXaTestCases CtsNetSecConfigInvalidPinTestCases CtsNetSecConfigNestedDomainConfigTestCases CtsSharedUserMigrationTestCases CtsPackageInstallAppOpDefaultTestCases CtsUidIsolationTestCases CtsNativeNetPlatformTestCases CtsTaskFpsCallbackTestCases CtsBroadcastTestCases CtsSharesheetTestCases CtsBackgroundRestrictionsTestCases CtsNetSecConfigDownloadManagerTestCases CtsNetSecConfigPrePCleartextTrafficTestCases CtsNetTestCasesLegacyApi22 CtsNetTestCasesLegacyPermission22 CtsSuspendAppsPermissionTestCases CtsSystemApiSignatureTestCases CtsFragmentTestCasesSdk26 CtsNetSecConfigAttributeTestCases CtsNetSecConfigBasicDebugDisabledTestCases CtsNetSecConfigBasicDebugEnabledTestCases CtsNoPermissionTestCases
Change-Id: Id0cf711b971da7004cd861b55a44f2c4c6eb96b8