Ensure events are received by window when emulating input device
Before this CL, the test would inject the events into the device, but
not check that any entity was actually getting the events. The goal was
for the metrics to be logged, and the test was waiting for the metric to
contain the correct number of events.

However, it's not clear at which point the input device should be
cleaned up. With that approach, the injection was happening in the
device code, but the metrics were getting observed from the test code.
It was difficult for the injector to understand the lifetime of the
caller, since the test device ideally should not be cleaned up until
the wait is over in the caller.

To get around this, use a different strategy for injection. In this CL,
we are going to actually look at the events that are received by the
test activity. This helps us establish an intermediate point where the
events should be processed, and thus make the virtual input device
lifetime more deterministic. In this CL, we can just wait for the
activity to receive the events, which would imply that the event has
been processed by the system, and therefore should now also be contained
in the metrics.

This allows us to eliminate things like Thread.sleep, and helps debug
future failures because we can trace the injected events more precisely.

Bug: 353332922
Test: atest CtsInputHostTestCases:InputAtomsTest
Flag: TEST_ONLY

Change-Id: I65201666e9ee7e77924c391166c76aff2cd7b0b2