Enable secondary_user_on_secondary_display for CtsApacheHttpTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264013

Test: atest --user-type secondary_user_on_secondary_display \
          CtsApacheHttpTestCases
Test: atest CtsApacheHttpTestCases
Change-Id: I190a5df55c0f0f0bc56ef155fb9dd5209ad2bc20