Enable secondary_user_on_secondary_display for CtsSearchUiServiceTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264986

Test: atest --user-type secondary_user_on_secondary_display \
          CtsSearchUiServiceTestCases
Test: atest CtsSearchUiServiceTestCases
Change-Id: I7c3ab67c3aba4de74591315c826fd8d95cbcdced