Update cts test rotation assumption
AttachedSurfaceControlTest assumes the system supports rotation when a
shell command returns "enabled". However this isn't always the case such
as in automotive where fixed-to-user-rotation is enabled by default this
causing the shell command check to fail when it returns "default. This
change updates the test to check to determine whether it can rotate or
not.

Bug: 355052385
Test: atest android.view.surfacecontrol.cts.AttachedSurfaceControlTest
Flag: TEST_ONLY
Change-Id: I8ca1410de35d58f9ba6199e517fac71d5b66ccc2