InputAtomsTest: Return touchpad swipes to the starting location
When interacting with the emulated touchpad, multi-finger swipes can
have unintended side effects, like expanding the notification shade,
which can alter the state of the system.

To combat this, return the swiping fingers to the starting location when
performing a multi-finger swipe, which should mitigate these side
effects.

Bug: 355035531
Test: atest CtsInputHostTestCases
Flag: TEST_ONLY
Change-Id: I1e80e73d666142dbce49cbdf1b3a3d0fd3fe6d9e