(3/n) Enable CTS test for secondary_user_on_secondary_display in bulk
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

Tested on gcar_md emu

Bug: 354233217
Test: atest --user-type secondary_user_on_secondary_display vm-tests-tf CtsNdkBinderTestCases CtsMediaEncoderTestCases CtsJdwpTestCases CtsUtilTestCases CtsResourcesLoaderTests CtsDatabaseTestCases CtsProtoTestCases CtsNfcTestCases CtsSensorRatePermissionTestCases CtsContentCaptureServiceTestCases CtsTfliteNnapiDelegateTestCases

Change-Id: Ie98038540c7f2005bb027a8ad6446499722fb3d7