Provide view-based hermetic injection functionality
For most of the use cases inside FocusHandlingTest, we can inject input
using in-process mechanisms like dispatchTouchEvent. This will keep the test fast and
deterministic, without affecting the entire system.

However, in some cases, it is necessary to indeed send a click event
through the entire system. For those situations, we introduce a
View-based injection to UinputTouchDevice. This will be more convenient
to use, and will keep the coordinate calculation inside the input-owned
libraries.

In some cases, I found that it wasn't needed to send any click events to
the test at all. The test still passes. For those cases, the "click"
emulation was removed completely.

Bug: 329439551
Flag: TEST_ONLY
Test: atest CtsInputMethodTestCases:FocusHandlingTest
Change-Id: Iaaa117026559f52aba91b4ed0ca2f089507a9c01