Always Attempt Reset of Bands When the TC Fails
There are a variety of "abnormal" failure cases, such as
when a binder call times out, or if an ISE is thrown that
will cause the testSetSystemSelectionChannels() to fail out
and exit ungracefully. In order to give the device the best
chance of recovering, this change will always attempt to
perform one last setSystemSelectionChannels() to the original
values. In this case, even if a queue of operations is still
pending, the final state has a better chance of being reasonable.

In the event that this operation itself also fails, the result
is no worse.

Bug: 354119793
Test: atest TelephonyManagerTest#testSetSystemSelectionChannels
Change-Id: Ic0cfea6f556cb465f7bf249757fc3ec2dd4bb8a2