Ignore telephony tests that are not relevant to SDK sandbox
Several CTS telephony tests cannot run in the SDK sandbox as the sandbox
package is not allowed to perform certain operations (because the
specific permissions are not declared in the sandbox package, the
sandbox does not have access to specific content providers or services
on the system etc.).

These test are ignroed using the AppModeNonSdkSandbox annotation. This
only affects tests when they are run in the sandbox mode, not when run
as usual.

Bug: 323156668
Flag: TEST_ONLY
Test: atest <test> -- --enable-optional-parameterization --enable-parameterized-modules --module-parameter run_on_sdk_sandbox
Change-Id: Iaa0c749407b60804665d533a67992bbac3713375