CtsPrintTestCases: Wait for session to end before starting another
After a print job is submitted, the activity is closed and the existing
sessions will be destroyed in the background.  If a new print job is
started quickly enough, it will reuse the existing session instead of
starting a new one.

The getActivePrintJobs test starts a new job without waiting for
teardown to finish, but then expects to see a new discovery session
start.  Close the race here by waiting for all of the previous discovery
sessions to be destroyed before starting a new print job.

Bug: 356221362
Test: atest CtsPrintTestCases:PrintServicesTest
Flag: TEST_ONLY
Change-Id: Idf4f9b290b077821b05371ddc8ffdb400c03f8e6