Enable secondary_user_on_secondary_display for CtsWindowManagerDeviceWindow
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 349704000
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceWindow
Test: atest --user-type secondary_user  CtsWindowManagerDeviceWindow
Test: atest CtsWindowManagerDeviceWindow
Flag: TEST_ONLY
Change-Id: I2d34a743788bdcd7f8248cc93f1458b18db58307