Skip 10-bit tests if device doesn't support P010
P010 support was added in Android T and requires HAL to also
support P010, so vendor image must also be at least T.

For before-T VNDK we need to also verify if YUVP010 color format is supported. In general for 10-bit HEIF decoding and 10-bit output we need the following restrictions to be fulfilled:
(1) has 10-bit HEVC decoder
(2) has P010 color format for that decoder (supported in T and after)

If only (1) is fulfilled the decoder will automatically fall back to outputting 8-bit color format and thus in this case we need to verify 8-bit output instead of 10-bit.

Bug: b/356536704
Change-Id: I42c20e02761727a913f27a71d0c99f702f46a557
Test: ImageDecoderTest#testDecode10BitHeif, BitmapFactoryTest#testDecode10BitHEIF10BitBitmap, BitmapFactoryTest#testDecode10BitHEIFTo8BitBitmap